// src/app/mechanic/models/car.model.ts
export interface Car {
  id?: number;
  matricula?: string;
  matricula_visible?: boolean;
  marca?: string;
  marca_visible?: boolean;
  modelo?: string;
  modelo_visible?: boolean;
  variante?: string;
  variante_visible?: boolean;
  tipo?: string;
  tipo_visible?: boolean;
  version?: string;
  version_visible?: boolean;
  primera_matriculacion?: string;
  primera_matriculacion_visible?: boolean;
  numero_bastidor?: string;
  numero_bastidor_visible?: boolean;
  combustible?: string;
  combustible_visible?: boolean;
  marca_motor?: string;
  marca_motor_visible?: boolean;
  codigo_motor?: string;
  codigo_motor_visible?: boolean;
  numero_cilindros?: string;
  numero_cilindros_visible?: boolean;
  cilindrada?: string;
  cilindrada_visible?: boolean;
  potencia_kw?: string;
  potencia_kw_visible?: boolean;
  potencia_cv?: string;
  potencia_cv_visible?: boolean;
  neumaticos?: string;
  neumaticos_visible?: boolean;
  categoria?: string;
  categoria_visible?: boolean;
  nivel_emisiones?: string;
  nivel_emisiones_visible?: boolean;
  codigo_radio?: null;
  codigo_radio_visible?: boolean;
  codigo_llaves?: null;
  codigo_llaves_visible?: boolean;
  reset_mantenimiento?: string;
  reset_mantenimiento_visible?: boolean;
  observaciones?: string;
  observaciones_visible?: boolean;
  created_at?: string;
  updated_at?: string;
}